requirements.txt
streamlit
diffusers
transformers
torch
accelerate
safetensors
Pillow
